+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.2.7 - 27 Dec 2024
 Bug fix:
 - macro doesn't work with guard
+- [#981](https://github.com/elysiajs/elysia/issues/981) unable to deference schema, create default, and coerce value
 
 # 1.2.6 - 25 Dec 2024
 Bug fix: